requirements.txt
Appium-Python-Client~=2.2.0
selenium~=4.1.3
openpyxl~=3.0.9
playsound~=1.2.2
pytz~=2022.1